// apps/ui/src/lib/api-client/models/OrganizationMember.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Proxy Smart Backend
 * SMART on FHIR Proxy + Healthcare Administration API using Keycloak and Elysia
 *
 * The version of the OpenAPI document: 0.0.5-alpha.202604151658.c3071eb4
 *
 * NOTE: This class is auto generated by generate-ts-fetch-client.py
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';

/**
 * @export
 * @interface OrganizationMember
 */
export interface OrganizationMember {
    /**
     * @type {string}
     * @memberof OrganizationMember
     */
    id?: string;
    /**
     * @type {string}
     * @memberof OrganizationMember
     */
    username?: string;
    /**
     * @type {string}
     * @memberof OrganizationMember
     */
    email?: string;
    /**
     * @type {string}
     * @memberof OrganizationMember
     */
    firstName?: string;
    /**
     * @type {string}
     * @memberof OrganizationMember
     */
    lastName?: string;
    /**
     * @type {boolean}
     * @memberof OrganizationMember
     */
    enabled?: boolean;
    /**
     * MANAGED or UNMANAGED
     * @type {string}
     * @memberof OrganizationMember
     */
    membershipType?: string;
}

/**
 * Check if a given object implements the OrganizationMember interface.
 */
export function instanceOfOrganizationMember(value: object): value is OrganizationMember {
    return true;
}

export function OrganizationMemberFromJSON(json: any): OrganizationMember {
    return OrganizationMemberFromJSONTyped(json, false);
}

export function OrganizationMemberFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrganizationMember {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'username': json['username'] == null ? undefined : json['username'],
        'email': json['email'] == null ? undefined : json['email'],
        'firstName': json['firstName'] == null ? undefined : json['firstName'],
        'lastName': json['lastName'] == null ? undefined : json['lastName'],
        'enabled': json['enabled'] == null ? undefined : json['enabled'],
        'membershipType': json['membershipType'] == null ? undefined : json['membershipType'],
    };
}

export function OrganizationMemberToJSON(json: any): OrganizationMember {
    return OrganizationMemberToJSONTyped(json, false);
}

export function OrganizationMemberToJSONTyped(value?: OrganizationMember | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'username': value['username'],
        'email': value['email'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'enabled': value['enabled'],
        'membershipType': value['membershipType'],
    };
}